chore: move wnodl and its peripherals to their own module under runtimes/main
